media-source/SourceBuffer-appendWindowEnd_updating.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>Getting SourceBuffer#appendWindowEnd during the updating attribute equals true</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
var contents = {'/media/white.webm' : 'video/webm; codecs="vorbis,vp8"',
                '/media/segmented-3s.mp4'  : 'video/mp4'};

//check the browser supports the MIME used in this test
function isTypeSupported(mime) {
    if(!MediaSource.isTypeSupported(mime)) {
        this.step(function() {
            assert_unreached("Browser doesn't support the MIME used in this test: " + mime);
        });
        this.done();
        return false;
    }
    return true;
}
function GET(url, processBody) {
    var xhr = new XMLHttpRequest();
    xhr.open('GET', url, true);
    xhr.responseType = 'arraybuffer';
    xhr.send();
    xhr.onload = function(e) {
        if (xhr.status != 200) {
            alert("Unexpected status code " + xhr.status + " for " + url);
            return false;
        }
        processBody(new Uint8Array(xhr.response));
    };
}
function test(file, mime) {
    async_test(function(t) {
        if(!isTypeSupported.bind(t)(mime)) {
            return;
        }
        GET(file, function(data) {
            var mediaSource = new MediaSource();
            mediaSource.addEventListener('sourceopen', function(e) {
                var sourceBuffer = mediaSource.addSourceBuffer(mime);
                sourceBuffer.addEventListener('updatestart', function(e) {
                    t.step(function() {
                        assert_throws({name: 'InvalidStateError'},
                                      function() {
                                          sourceBuffer.appendWindowEnd = 1;
                                      });
                    });
                    t.done();
                });
                sourceBuffer.appendBuffer(data);
            });
            video.src = window.URL.createObjectURL(mediaSource);
        });
    }, 'SourceBuffer#appendWindowEnd (' + mime + '): If the updating attribute equals true, INVALID_STATE_ERR exception should be thrown.');
}
for(var filename in contents) {
    test(filename, contents[filename]);
}
</script>
</body>
</html>
